fixup! refactor(compiler): integrate let declarations into the template pipeline
