+++ packages/compiler/src/template/pipeline/src/phases/resolve_names.ts
@@ -61,9 +61,6 @@ function processLexicalScope(
             break;
         }
         break;
-      case ir.OpKind.StoreLet:
-        scope.set(op.declaredName, op.target);
-        break;
       case ir.OpKind.Listener:
       case ir.OpKind.TwoWayListener:
         // Listener functions have separate variable declarations, so process them as a separate
